requirements/base.txt
Django==2.1.2
psycopg2==2.7.5
pytz==2018.5
# Django Rest Framework
djangorestframework==3.8.2
# Django Treebeard
django-treebeard==4.3
# LDAP 3
ldap3==2.5.1
pyasn1==0.4.4
# Django Reversion
django-reversion==3.0.1